+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ## 2.3.0 (Released 2023-09-xx)
 
 - Added an exit confirmation if the project has unsaved changes.
+- Improved the phrase generator for arrangements with existing DD levels.
 - Fixed a possible false positive with the HOPO into same fret validation after a slide.
 - Fixed an issue in the base tone automatic naming.